+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- VMEET 2218 - Add tab private chat and tab general chat
+
 ### Fixed
 - MV255 - move chat css to event guideline css
 - MV255 - move question css to event guideline css